+++ src/Modeles/entities/Entity3D.java
@@ -1,9 +1,9 @@
 package Modeles.entities;
 
 public interface Entity3D {
-	public float getX();
-	public float getY();
-	public float getZ();
+	public int getX();
+	public int getY();
+	public int getZ();
 	public void setX(float x);
 	public void setY(float y);
 	public void setz(float z);
